requirements.txt
numpy
pillow
scikit-learn
matplotlib
joblib